requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.29.0
openai==1.82.0
python-dotenv==1.0.1
pydantic==2.11.5
python-multipart==0.0.20
pynamodb==6.1.0
boto3==1.40.40
celery==5.3.4
redis==5.0.1
slowapi==0.1.9
langchain-community==0.3.27
langchain-openai==0.2.11
langchain-core==0.3.66
docx2txt==0.9
pypdf==6.1.0
requests==2.32.3
numpy>=2.1.0,<3.0.0
psycopg2-binary==2.9.10
pgvector==0.2.4
sqlalchemy==2.0.36
alembic==1.14.0
sentence-transformers==3.2.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.24.0
cachetools==5.5.0